fix: refactor fee handling by consolidating FeeLabel component and updating references for improved clarity
